requirements.txt
streamlit==1.31.1
google-generativeai==0.3.2
Pillow==10.2.0
PyPDF2==3.0.1
python-docx==0.8.11
